Fix Metal training by using f32 tensors instead of f64
Metal's index_select kernel doesn't support the U32×F64 combination,
causing "Metal contiguous index_select U32 F64 not implemented" at
training time. Switch all candle tensor creation to f32 while keeping
the JSON weight format compatible (f64) for InferenceGpt interop.

Bump microgpt CLI to 0.2.1.

+++ domains/ai/libs/microgpt/src/tensor_train.rs
@@ -57,7 +57,7 @@ pub fn tensor_train_step(
     // Cross-entropy loss
     let targets = Tensor::new(target_tokens, &model.device)?;
     let loss = candle_nn::loss::cross_entropy(&logits, &targets)?;
-    let loss_val = loss.to_scalar::<f64>()?;
+    let loss_val = loss.to_scalar::<f32>()? as f64;
 
     // Backward + update
     optimizer.inner.backward_step(&loss)?;
